# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "AMIDDOCS OFFICIAL SWAGGER",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/departments/create": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    },
                    {
                        "Token": []
                    }
                ],
                "description": "enabled auto check unique name and shortName values",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "departments"
                ],
                "summary": "Create department",
                "parameters": [
                    {
                        "description": "department dto",
                        "name": "department",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/depmodel.CreateDepartmentDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/depmodel.DepartmentDTO"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/departments/get-by-id": {
            "get": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "return department dto by id in param",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "departments"
                ],
                "summary": "Get Department By own ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "department id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/depmodel.DepartmentDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/document-templates/get": {
            "get": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "returns raw file",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "document-templates"
                ],
                "summary": "get template document",
                "parameters": [
                    {
                        "enum": [
                            "STUDY_DOCUMENT_BUDGET",
                            "STUDY_DOCUMENT_NO_BUDGET"
                        ],
                        "type": "string",
                        "description": "document type",
                        "name": "type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "department id",
                        "name": "departmentId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/document-templates/upload": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    },
                    {
                        "Token": []
                    }
                ],
                "description": "load or update template in files",
                "consumes": [
                    "application/octet-stream"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "document-templates"
                ],
                "summary": "load document template",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "department id",
                        "name": "depId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "STUDY_DOCUMENT_BUDGET",
                            "STUDY_DOCUMENT_NO_BUDGET"
                        ],
                        "type": "string",
                        "description": "document type",
                        "name": "type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "document binary (.docx file)",
                        "name": "document",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/groups/create": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    },
                    {
                        "Token": []
                    }
                ],
                "description": "create group",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "CreateGroup",
                "parameters": [
                    {
                        "description": "group dto",
                        "name": "groups",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/groupmodel.GroupDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/groupmodel.GroupDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/groups/get-by-id": {
            "get": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "Get GroupDTO by Id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Get Group By Id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "group id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/groupmodel.GroupDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "amiderrors.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "depmodel.CreateDepartmentDTO": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "shortName": {
                    "type": "string"
                }
            }
        },
        "depmodel.DepartmentDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "imageUrl": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "groupfields.EducationForm": {
            "type": "string",
            "enum": [
                "FULL_TIME",
                "EXTRAMURAL"
            ],
            "x-enum-varnames": [
                "FULL_TIME",
                "EXTRAMURAL"
            ]
        },
        "groupmodel.GroupDTO": {
            "type": "object",
            "properties": {
                "educationFinishDate": {
                    "type": "string"
                },
                "educationForm": {
                    "$ref": "#/definitions/groupfields.EducationForm"
                },
                "educationStartDate": {
                    "type": "string"
                },
                "educationYear": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "isBudget": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "studyDepartmentId": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Bearer Token Auth",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "Token": {
            "description": "Client Token auth",
            "type": "apiKey",
            "name": "Token",
            "in": "header"
        }
    }
}